The meeting brief's evidence collector gets its own file
service.go crossed the 500-line cap. The collector is one concept — which
citations in a finished brief could name a message — and it moves out whole,
the way org360's does.
